requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.8
django-3-jet==1.0.8
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==9.7.1
google-api-python-client==1.4.1
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.3
oauth2client==4.1.3
openapi-codec==1.3.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.0
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.3
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-selenium==2.0.1
pytest-variables==1.9.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
selenium==4.0.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tenacity==6.3.1
text-unidecode==1.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
wsproto==1.0.0
